Initial definition of common device API and CpuDevice class
Commit refactors device.h to define an initial version of a common device interface. Implementation of CPU device is updated to match the defined functions. Also, CPU implementation moves to a CpuDevice class derived from DeviceBase. In the future this should allow for better handling of various devices, as common features can be implemented in the DeviceBase.

Signed-off-by: Lukasz Tuz <lukasz.tuz@intel.com>